+++ lib/src/data/converter/community_hive_entity_extension_converter.dart
@@ -20,6 +20,7 @@ extension CommunityHiveEntityExtension on CommunityHiveEntity {
       ..isDeleted = isDeleted
       ..createdAt = createdAt
       ..updatedAt = editedAt
+      ..allowCommentInStory = allowCommentInStory
       ..categoryIds = categoryIds
       ..isPostReviewEnabled = needApprovalOnPostCreation
       ..tags = tags