+++ lib/src/data/data_source/remote/api_interface/network_settings_api_interface.dart
@@ -1,0 +1,9 @@
+import 'package:amity_sdk/src/core/model/api_request/validate_texts_request.dart';
+import 'package:amity_sdk/src/core/model/api_request/validate_urls_request.dart';
+
+abstract class NetworkSettingsApiInterface{
+  
+  Future<bool> validateUrls(ValidateUrlsRequest request);
+  Future<bool> validateTexts(ValidateTextsRequest request);
+
+}