+++ lib/src/data/data_source/remote/api_interface/story_target_api_interface.dart
@@ -1,7 +1,9 @@
+import 'package:amity_sdk/src/core/model/api_request/get_global_story_target_request.dart';
 import 'package:amity_sdk/src/core/model/api_request/get_targets_by_targets_request.dart';
-import 'package:amity_sdk/src/data/response/create_story_response.dart';
+import 'package:amity_sdk/src/data/response/get_global_story_target_response.dart';
 import 'package:amity_sdk/src/data/response/get_story_targets_by_targets_response.dart';
 
 abstract class StoryTargetApiInterface{
   Future<GetStoryTargetsByTargets> queryStoryTargets( GetTargetsByTargetsRequest request);
+  Future<GetGlobalStoryTargetResponse> queryGlobalFeed( GetGlobalStoryTargetRequest request);
 }