+++ lib/src/domain/model/amity_notification_settings/amity_notification.dart
@@ -1,0 +1,14 @@
+
+import 'package:amity_sdk/amity_sdk.dart';
+
+class AmityNotification{
+
+  AmityCommunityNotification community(String communityId) {
+        return AmityCommunityNotification( communityId);
+    }
+
+
+
+
+  
+}