+++ lib/src/domain/repo/story_target_repo.dart
@@ -1,7 +1,10 @@
 import 'dart:async';
 
 import 'package:amity_sdk/amity_sdk.dart';
+import 'package:amity_sdk/src/core/core.dart';
+import 'package:amity_sdk/src/core/model/api_request/get_global_story_target_request.dart';
 import 'package:amity_sdk/src/core/model/api_request/get_targets_by_targets_request.dart';
+import 'package:amity_sdk/src/core/model/api_request/stream_query_request.dart';
 import 'package:amity_sdk/src/data/data_source/local/hive_entity/story_target_hive_entity_28.dart';
 import 'package:amity_sdk/src/domain/repo/amity_object_repository.dart';
 
@@ -16,6 +19,8 @@ abstract class StoryTargetRepo
 
   Stream<List<AmityStoryTarget>> listenStoryTargets(
       RequestBuilder<GetTargetsByTargetsRequest> request);
+
+  Stream<List<AmityStoryTarget>> listenAllStoryTargets( );
 
   Future updateStoryTargetLocalLastStoryExpiresAt(
     AmityStoryTargetType targetType,
@@ -32,5 +37,7 @@ abstract class StoryTargetRepo
       AmityStoryTargetType targetType,
       String targetId,
       DateTime? lastStoryExpiresAt);
+
+  Future<PageListData<List<AmityStoryTarget>, String>> getGlobalFeed( GetGlobalStoryTargetRequest request);
 
 }